third_party/WebKit/LayoutTests/platform/win/external/wpt/html/browsers/windows/nested-browsing-contexts/window-top-null-expected.txt

This is a testharness.js-based test.
PASS `window.top` is null when browsing context container element removed 
PASS `window.top`null when any ancestor browsing context container removed 
Harness: the test ran to completion.
